+++ source/common/formatter/BUILD
@@ -57,3 +57,62 @@ envoy_cc_library(
         "//source/common/stream_info:utility_lib",
     ],
 )
+
+# The library is compiled as extension and linked to the main library.
+envoy_cc_library(
+    name = "http_speicific_formatter_lib",
+    srcs = ["http_specific_formatter.cc"],
+    hdrs = ["http_specific_formatter.h"],
+    # This visibility is used to ensure this library will works like an extension
+    # and could only be linked to the main library (or test library).
+    visibility = [
+        "//source/exe:__subpackages__",
+        "//test:__subpackages__",
+    ],
+    deps = [
+        "//envoy/api:api_interface",
+        "//envoy/formatter:substitution_formatter_interface",
+        "//envoy/runtime:runtime_interface",
+        "//envoy/stream_info:stream_info_interface",
+        "//envoy/upstream:upstream_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/common:utility_lib",
+        "//source/common/formatter:substitution_format_utility_lib",
+        "//source/common/grpc:common_lib",
+        "//source/common/http:utility_lib",
+        "//source/common/json:json_loader_lib",
+        "//source/common/protobuf:message_validator_lib",
+        "//source/common/stream_info:utility_lib",
+    ],
+    alwayslink = True,
+)
+
+# The library is compiled as extension and linked to the main library.
+envoy_cc_library(
+    name = "stream_info_formatter_lib",
+    srcs = ["stream_info_formatter.cc"],
+    hdrs = ["stream_info_formatter.h"],
+    # This visibility is used to ensure this library will works like an extension
+    # and could only be linked to the main library (or test library).
+    visibility = [
+        "//source/exe:__subpackages__",
+        "//test:__subpackages__",
+    ],
+    deps = [
+        "//envoy/api:api_interface",
+        "//envoy/formatter:substitution_formatter_interface",
+        "//envoy/runtime:runtime_interface",
+        "//envoy/stream_info:stream_info_interface",
+        "//envoy/upstream:upstream_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/common:random_generator_lib",
+        "//source/common/common:utility_lib",
+        "//source/common/formatter:substitution_format_utility_lib",
+        "//source/common/grpc:common_lib",
+        "//source/common/http:utility_lib",
+        "//source/common/json:json_loader_lib",
+        "//source/common/protobuf:message_validator_lib",
+        "//source/common/stream_info:utility_lib",
+    ],
+    alwayslink = True,
+)
