+++ source/common/formatter/stream_info_formatter.h
