+++ test/common/formatter/substitution_formatter_test.cc
@@ -9,6 +9,8 @@
 
 #include "source/common/common/logger.h"
 #include "source/common/common/utility.h"
+#include "source/common/formatter/http_specific_formatter.h"
+#include "source/common/formatter/stream_info_formatter.h"
 #include "source/common/formatter/substitution_format_utility.h"
 #include "source/common/formatter/substitution_formatter.h"
 #include "source/common/http/header_map_impl.h"
@@ -17,8 +19,6 @@
 #include "source/common/protobuf/utility.h"
 #include "source/common/router/string_accessor_impl.h"
 #include "source/common/stream_info/stream_id_provider_impl.h"
-#include "source/extensions/formatter/built_in_http/http_specific_formatter.h"
-#include "source/extensions/formatter/built_in_stream_info/stream_info_formatter.h"
 
 #include "test/common/formatter/command_extension.h"
 #include "test/mocks/api/mocks.h"
