+++ test/common/http/BUILD
@@ -239,13 +239,13 @@ envoy_cc_test_library(
     hdrs = ["conn_manager_impl_test_base.h"],
     deps = [
         ":xff_extension_lib",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/http:conn_manager_lib",
         "//source/common/http:context_lib",
         "//source/common/http:rds_lib",
         "//source/common/tracing:custom_tag_lib",
         "//source/extensions/access_loggers/common:file_access_log_lib",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//source/extensions/http/header_validators/envoy_default:http1_header_validator",
         "//source/extensions/request_id/uuid:config",
         "//test/mocks/access_log:access_log_mocks",
@@ -287,13 +287,13 @@ envoy_cc_test(
         ":xff_extension_lib",
         "//source/common/common:random_generator_lib",
         "//source/common/event:dispatcher_lib",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/http:conn_manager_lib",
         "//source/common/http:headers_lib",
         "//source/common/network:address_lib",
         "//source/common/network:utility_lib",
         "//source/common/runtime:runtime_lib",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//source/extensions/request_id/uuid:config",
         "//test/mocks/http:http_mocks",
         "//test/mocks/local_info:local_info_mocks",
@@ -588,9 +588,9 @@ envoy_cc_test(
     name = "header_mutation_test",
     srcs = ["header_mutation_test.cc"],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/http:header_mutation_lib",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/mocks/stream_info:stream_info_mocks",
         "//test/test_common:utility_lib",
     ],
