+++ test/extensions/access_loggers/open_telemetry/BUILD
@@ -37,11 +37,11 @@ envoy_extension_cc_test(
     extension_names = ["envoy.access_loggers.open_telemetry"],
     deps = [
         "//envoy/common:time_interface",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/protobuf",
         "//source/common/router:string_accessor_lib",
         "//source/extensions/access_loggers/open_telemetry:access_log_lib",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/mocks/access_log:access_log_mocks",
         "//test/mocks/grpc:grpc_mocks",
         "//test/mocks/local_info:local_info_mocks",
@@ -61,10 +61,10 @@ envoy_extension_cc_test(
     srcs = ["config_test.cc"],
     extension_names = ["envoy.access_loggers.open_telemetry"],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/extensions/access_loggers/open_telemetry:access_log_lib",
         "//source/extensions/access_loggers/open_telemetry:config",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/mocks/server:factory_context_mocks",
         "@envoy_api//envoy/extensions/access_loggers/grpc/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/access_loggers/open_telemetry/v3:pkg_cc_proto",
@@ -78,11 +78,11 @@ envoy_extension_cc_test(
     extension_names = ["envoy.access_loggers.open_telemetry"],
     deps = [
         "//source/common/buffer:zero_copy_input_stream_lib",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/grpc:codec_lib",
         "//source/common/grpc:common_lib",
         "//source/extensions/access_loggers/open_telemetry:config",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/common/grpc:grpc_client_integration_lib",
         "//test/integration:http_integration_lib",
         "//test/test_common:utility_lib",
@@ -106,14 +106,14 @@ envoy_extension_cc_test(
     extension_names = ["envoy.access_loggers.open_telemetry"],
     tags = ["skip_on_windows"],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/formatter:substitution_formatter_lib",
         "//source/common/http:exception_lib",
         "//source/common/http:header_map_lib",
         "//source/common/router:string_accessor_lib",
         "//source/extensions/access_loggers/open_telemetry:config",
         "//source/extensions/access_loggers/open_telemetry:substitution_formatter_lib",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//source/extensions/formatter/cel:cel_lib",
         "//source/extensions/formatter/cel:config",
         "//test/mocks/server:factory_context_mocks",
@@ -138,11 +138,11 @@ envoy_cc_benchmark_binary(
         "benchmark",
     ],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/http:header_map_lib",
         "//source/common/network:address_lib",
         "//source/extensions/access_loggers/open_telemetry:substitution_formatter_lib",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/common/stream_info:test_util",
         "//test/mocks/stream_info:stream_info_mocks",
         "@opentelemetry_proto//:common_cc_proto",
