+++ test/extensions/access_loggers/stream/BUILD
@@ -14,8 +14,8 @@ envoy_cc_test(
     srcs = ["stdout_config_test.cc"],
     deps = [
         ":stream_test_base",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "@envoy_api//envoy/extensions/access_loggers/stream/v3:pkg_cc_proto",
     ],
 )
@@ -25,8 +25,8 @@ envoy_cc_test(
     srcs = ["stderr_config_test.cc"],
     deps = [
         ":stream_test_base",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "@envoy_api//envoy/extensions/access_loggers/stream/v3:pkg_cc_proto",
     ],
 )
@@ -35,9 +35,9 @@ envoy_cc_test_library(
     name = "stream_test_base",
     hdrs = ["stream_test_base.h"],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/extensions/access_loggers/stream:config",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/mocks/server:factory_context_mocks",
         "//test/test_common:environment_lib",
         "//test/test_common:utility_lib",
