+++ test/extensions/filters/http/header_mutation/BUILD
@@ -18,9 +18,9 @@ envoy_extension_cc_test(
     ],
     extension_names = ["envoy.filters.http.header_mutation"],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/extensions/filters/http/header_mutation:config",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/mocks/api:api_mocks",
         "//test/mocks/http:http_mocks",
         "//test/mocks/server:factory_context_mocks",
@@ -34,9 +34,9 @@ envoy_extension_cc_test(
     srcs = ["config_test.cc"],
     extension_names = ["envoy.filters.http.header_mutation"],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/extensions/filters/http/header_mutation:config",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/mocks/server:factory_context_mocks",
         "//test/mocks/server:instance_mocks",
         "//test/test_common:registry_lib",
@@ -50,9 +50,9 @@ envoy_extension_cc_test(
     srcs = ["header_mutation_integration_test.cc"],
     extension_names = ["envoy.filters.http.header_mutation"],
     deps = [
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/extensions/filters/http/header_mutation:config",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//test/integration:http_integration_lib",
         "//test/mocks/server:instance_mocks",
         "//test/test_common:registry_lib",
