+++ test/server/admin/BUILD
@@ -36,8 +36,8 @@ envoy_cc_test_library(
     srcs = envoy_select_admin_functionality(["admin_instance.cc"]),
     hdrs = envoy_select_admin_functionality(["admin_instance.h"]),
     deps = [
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/server:configuration_lib",
         "//source/server/admin:admin_lib",
         "//test/mocks/runtime:runtime_mocks",
@@ -55,13 +55,13 @@ envoy_cc_test(
         ":admin_instance_lib",
         "//envoy/json:json_object_interface",
         "//envoy/runtime:runtime_interface",
+        "//source/common/formatter:http_speicific_formatter_lib",
+        "//source/common/formatter:stream_info_formatter_lib",
         "//source/common/http:message_lib",
         "//source/common/json:json_loader_lib",
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
         "//source/common/stats:thread_local_store_lib",
-        "//source/extensions/formatter/built_in_http:config",
-        "//source/extensions/formatter/built_in_stream_info:config",
         "//source/server:null_overload_manager_lib",
         "//source/server/admin:admin_lib",
         "//test/mocks/runtime:runtime_mocks",
